vktrace trim: int -> size_t
XCAP-536

+++ vktrace/vktrace_layer/vktrace_lib_trim.cpp
@@ -1588,7 +1588,7 @@ void remove_Device_object(VkDevice var) {
         }
     }
 
-    for (int i = 0; i < queuesToRemove.size(); i++)
+    for (size_t i = 0; i < queuesToRemove.size(); i++)
     {
         trim::remove_Queue_object(queuesToRemove[i]);
     }
